OffloadedSNARK: move to module `snark`
